+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vue-vine/eslint-config
 
+## 1.1.11
+
+### Patch Changes
+
+- Fix typeof guard expression not complete for #309
+- Updated dependencies
+  - @vue-vine/eslint-plugin@1.1.11
+
+## 1.1.10
+
+### Patch Changes
+
+- Fix HMR ts-morph add source file panic and unrecongnized reference for template TS code in ESLint parser.
+- Updated dependencies
+  - @vue-vine/eslint-plugin@1.1.10
+
 ## 1.1.9
 
 ### Patch Changes