+++ packages/tsc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # vue-vine-tsc
 
+## 1.7.7
+
+### Patch Changes
+
+- Fix HMR ts-morph add source file panic and unrecongnized reference for template TS code in ESLint parser.
+- Updated dependencies
+  - @vue-vine/language-service@1.7.7
+
 ## 1.7.6
 
 ### Patch Changes