+++ packages/vue-vine/CHANGELOG.md
@@ -1,5 +1,13 @@
 # vue-vine
 
+## 1.7.7
+
+### Patch Changes
+
+- Fix HMR ts-morph add source file panic and unrecongnized reference for template TS code in ESLint parser.
+- Updated dependencies
+  - @vue-vine/vite-plugin@1.7.7
+
 ## 1.7.6
 
 ### Patch Changes